% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature.reduce.R
\name{Feature.reduce}
\alias{Feature.reduce}
\title{Reduce the feature dimension of the input data}
\usage{
Feature.reduce(X, Y = NULL, method = "GRM", dimension = "Ind.num")
}
\arguments{
\item{X}{genotype in numeric format, pure 0, 1, 2 matrix; geno is n individual rows by m snps columns}

\item{Y}{a phenotype vector of type numeric}

\item{method}{the method for reducing the feature; options are GRM, PCA, GWAS(MLM)}

\item{dimension}{confirm the dimensions after the reduction}
}
\value{
X data after dimensionality reduction
}
\description{
\code{Feature.reduce}Reduce the feature dimension of the input data
}
\examples{
data(example_dataset)
X <- genotype
X.FR <- Feature.reduce(X = X, method="GRM")

}
