% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\docType{data}
\name{genotype}
\alias{genotype}
\title{genotype}
\source{
R package GAPIT
}
\arguments{
\item{genotype}{a matrix}
}
\description{
genotype matrix which is example data from R package GAPIT
}
\keyword{datasets}
