% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\docType{data}
\name{phenotype}
\alias{phenotype}
\title{phenotype}
\source{
R package GAPIT
}
\arguments{
\item{phenotype}{a matrix}
}
\description{
phenotype matrix which is example data from R package GAPIT
}
\keyword{datasets}
