% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.env.R
\name{python.env}
\alias{python.env}
\title{Load the python environment, and import scikit-learn and numpy module}
\usage{
python.env(env = "python", env.path, module.check = TRUE)
}
\arguments{
\item{env}{the environment of python, including python, conda, miniconda}

\item{env.path}{the path of python environment}

\item{module.check}{check that the required modules are installed}
}
\value{
environment and module
}
\description{
\code{python.env}Load the python environment, and import scikit-learn and numpy module
}
\examples{
library(reticulate)
path <- reticulate::py_discover_config()
python.env(env="python", env.path=path$python, module.check=TRUE)

}
