% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.ML.R
\name{train.ML}
\alias{train.ML}
\title{Train single ML model and tune its parameters}
\usage{
train.ML(
  y_train,
  X_train,
  X_test,
  method,
  cv = 5,
  py.lib = py.lib,
  parameters = NULL
)
}
\arguments{
\item{y_train}{a phenotype vector of type numeric for training}

\item{X_train}{a genotype matrix of type numeric for training}

\item{X_test}{a genotype matrix of type numeric, the verification group(or no phenotype group) required to calculate GEBV}

\item{method}{a character, options are Support Vector Regression (SVR), Kernel Ridge Regression (KRR), Ridge Regression (Ridge)}

\item{cv}{the number of folds of cross-validation, the default is 5}

\item{py.lib}{load python module sklearn and numpy}

\item{parameters}{a list of user-specified parameters, the default is NULL}
}
\value{
A list of prediction values from single ML model and its optimal parameters
}
\description{
\code{train.ML}Train a single machine learning model and tune its parameters
}
\examples{
data(example_dataset)
X <- genotype
X.FR <- Feature.reduce(X = X, method="GRM")
y.train <- phenotype[c(1:200),1]
X.train <- X.FR[c(1:200),]
X.test <- X.FR[-c(1:200),]
path <- reticulate::py_discover_config()
python.env(env="python", env.path=path$python, module.check=TRUE)
py.lib <- python.env(env="python", env.path=path$python, module.check=TRUE)
SVR.fit <- train.ML(y_train=y.train,X_train=X.train,X_test=X.test,method="SVR",py.lib=py.lib)

}
