% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VanRaden.K.R
\name{VanRaden.K}
\alias{VanRaden.K}
\title{construct a kinship matrix by VanRaden method}
\usage{
VanRaden.K(geno, maf = NULL)
}
\arguments{
\item{geno}{a genotype matrix of type numeric}

\item{maf}{a maf is allele frequency of second allele vector}
}
\value{
a kinship matrix
}
\description{
\code{VanRaden.K}use VanRaden method to construct the kinship matrix
}
\examples{
data(example_dataset)
X <- genotype
G <- VanRaden.K(geno = X,maf=NULL)

}
